// src/interfaces/artworkInterfaces.ts
  export interface SimpleArtwork {
    id:                 number;
    title:              string;
    thumbnail:          Thumbnail;
    date_start:         number;
    date_display:       string;
    artist_display:     string;
    dimensions:         string;
    medium_display:     string;
    artwork_type_title: string;
    artist_titles:      string[];
    image_id:           string;
    has_not_been_viewed_much: boolean;
  }
  
  export interface Thumbnail {
    lqip:     string;
    width:    number;
    height:   number;
    alt_text: string;
  }